+++ src/main/java/com/tactfactory/capfakeskillspring/dao/interfaces/base/IBaseDAO.java
@@ -14,7 +14,7 @@ public abstract interface IBaseDAO<T extends BaseEntity> {
 
 	public List<T> getAll();
 
-	public T getById(Integer id);
+	public T getById(Long id);
 
 	public void update(T item);
 }
