+++ src/main/java/com/tactfactory/capfakeskillspring/models/base/BaseEntity.java
@@ -10,19 +10,19 @@ public class BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Integer id;
+	private Long id;
 
 	public BaseEntity() {
     }
-	public BaseEntity(Integer id) {
+	public BaseEntity(Long id) {
         this.setId(id);
     }
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
